moved web-app specific utils into app's module
