+++ python/smqtk/web/search_app/mongo_sessions.py
